ORY-49: move files to new 'internal' directory for clarity
