+++ authentication/src/lib/components/ory/internal/FlowInput.svelte
@@ -1,18 +1,18 @@
 <script lang="ts">
     import type { UiNode, UiNodeInputAttributes } from "@ory/client";
     import { UiNodeInputAttributesTypeEnum } from "@ory/client/api";
-    import FlowInputText from "$lib/components/ory/input/FlowInputText.svelte";
-    import FlowInputPassword from "$lib/components/ory/input/FlowInputPassword.svelte";
-    import FlowInputNumber from "$lib/components/ory/input/FlowInputNumber.svelte";
-    import FlowInputCheckbox from "$lib/components/ory/input/FlowInputCheckbox.svelte";
-    import FlowInputHidden from "$lib/components/ory/input/FlowInputHidden.svelte";
-    import FlowInputEmail from "$lib/components/ory/input/FlowInputEmail.svelte";
-    import FlowInputSubmit from "$lib/components/ory/input/FlowInputSubmit.svelte";
-    import FlowInputButton from "$lib/components/ory/input/FlowInputButton.svelte";
-    import FlowInputTel from "$lib/components/ory/input/FlowInputTel.svelte";
-    import FlowInputDateTimeLocal from "$lib/components/ory/input/FlowInputDateTimeLocal.svelte";
-    import FlowInputDate from "$lib/components/ory/input/FlowInputDate.svelte";
-    import FlowInputUrl from "$lib/components/ory/input/FlowInputUrl.svelte";
+    import FlowInputText from "$lib/components/ory/internal/input/FlowInputText.svelte";
+    import FlowInputPassword from "$lib/components/ory/internal/input/FlowInputPassword.svelte";
+    import FlowInputNumber from "$lib/components/ory/internal/input/FlowInputNumber.svelte";
+    import FlowInputCheckbox from "$lib/components/ory/internal/input/FlowInputCheckbox.svelte";
+    import FlowInputHidden from "$lib/components/ory/internal/input/FlowInputHidden.svelte";
+    import FlowInputEmail from "$lib/components/ory/internal/input/FlowInputEmail.svelte";
+    import FlowInputSubmit from "$lib/components/ory/internal/input/FlowInputSubmit.svelte";
+    import FlowInputButton from "$lib/components/ory/internal/input/FlowInputButton.svelte";
+    import FlowInputTel from "$lib/components/ory/internal/input/FlowInputTel.svelte";
+    import FlowInputDateTimeLocal from "$lib/components/ory/internal/input/FlowInputDateTimeLocal.svelte";
+    import FlowInputDate from "$lib/components/ory/internal/input/FlowInputDate.svelte";
+    import FlowInputUrl from "$lib/components/ory/internal/input/FlowInputUrl.svelte";
 
     export let node: UiNode;
     const attributes = node.attributes as UiNodeInputAttributes;
