+++ authentication/src/lib/components/ory/internal/FlowSet.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
     import type { UiNode } from "@ory/client";
     import { UiNodeTypeEnum } from "@ory/client";
-    import FlowImage from "$lib/components/ory/FlowImage.svelte";
-    import FlowText from "$lib/components/ory/FlowText.svelte";
-    import FlowInput from "$lib/components/ory/input/FlowInput.svelte";
+    import FlowImage from "$lib/components/ory/internal/FlowImage.svelte";
+    import FlowText from "$lib/components/ory/internal/FlowText.svelte";
+    import FlowInput from "$lib/components/ory/internal/input/FlowInput.svelte";
     import FlowScript from "./FlowScript.svelte";
     import FlowAnchor from "./FlowAnchor.svelte";
