+++ authentication/src/lib/components/ory/internal/input/FlowInputButton.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { t, fromNode } from "$lib/i18n";
+    import { fromNode, t } from "$lib/i18n";
     import type { UiNode, UiNodeInputAttributes } from "@ory/client";
 
     export let node: UiNode;
