+++ authentication/src/lib/components/ory/internal/input/FlowInputText.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { t, fromNode } from "$lib/i18n";
+    import { fromNode, t } from "$lib/i18n";
     import Messages from "$lib/components/ory/Messages.svelte";
     import type { UiNode, UiNodeInputAttributes } from "@ory/client";
 
@@ -16,7 +16,7 @@
             name={attributes.name}
             placeholder={$t(fromNode(node), node.meta.label?.context)}
             required={attributes.required}
-            type="number"
+            type="text"
             value={attributes.value ?? ""}
         />
     </label>
